sample-ecommerce-app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
httpx==0.27.0
pyjwt[crypto]==2.9.0
python-multipart==0.0.9
boto3==1.34.162

